+++ R/whisper.R
@@ -144,7 +144,12 @@ predict.whisper <- function(object, newdata, type = c("transcribe", "translate")
 #'                  language = "en", duration = 1000)
 #' }
 whisper <- function(x, use_gpu = FALSE, overwrite = FALSE, model_dir = Sys.getenv("WHISPER_MODEL_DIR", unset = getwd()), ...){
-  if(x %in% c("tiny", "tiny.en", "base", "base.en", "small", "small.en", "medium", "medium.en", "large-v1", "large-v2", "large-v3", "large")){
+  if(x %in% c("tiny", "tiny.en", "base", "base.en", "small", "small.en", "medium", "medium.en", "large-v1", "large-v2", "large-v3", "large",
+              "tiny-q5_1", "tiny.en-q5_1", 
+              "base-q5_1", "base.en-q5_1", 
+              "small-q5_1", "small.en-q5_1", 
+              "medium-q5_1", "medium.en-q5_1", 
+              "large-v2-q5_0", "large-v3-q5_0")){
     x <- whisper_download_model(x, overwrite = overwrite, model_dir = model_dir)
   }
   if(inherits(x, "whisper_download")){
@@ -166,9 +171,10 @@ whisper <- function(x, use_gpu = FALSE, overwrite = FALSE, model_dir = Sys.geten
 #' \item{small & small.en: 466 MB, RAM required: ~1.0 GB. Multilingual and English only version.}
 #' \item{medium & medium.en: 1.5 GB, RAM required: ~2.6 GB. Multilingual and English only version.}
 #' \item{large-v1, large-v2, large-v3: 2.9 GB, RAM required: ~4.7 GB. Multilingual}
+#' \item{quantised models: tiny-q5_1, tiny.en-q5_1, base-q5_1, base.en-q5_1, small-q5_1, small.en-q5_1, medium-q5_1, medium.en-q5_1, arge-v2-q5_0 and large-v3-q5_0 (only - from version 1.5.4 onwards)}
 #' }
 #' Note that the larger models may take longer than 60 seconds to download, so consider 
-#' increasing the timeout option in R via \code{options(timeout=120)}
+#' increasing the timeout option in R via \code{options(timeout = 120)}
 #' @param x the name of the model
 #' @param model_dir a path where the model will be downloaded to. Defaults to the environment variable \code{WHISPER_MODEL_DIR} and if this is not set, the current working directory
 #' @param repos character string with the repository to download the model from. Either
@@ -205,11 +211,22 @@ whisper <- function(x, use_gpu = FALSE, overwrite = FALSE, model_dir = Sys.geten
 #' whisper_download_model("large-v1")
 #' whisper_download_model("large-v2")
 #' whisper_download_model("large-v3")
+#' whisper_download_model("tiny-q5_1")
+#' whisper_download_model("base-q5_1")
+#' whisper_download_model("small-q5_1")
+#' whisper_download_model("medium-q5_1")
+#' whisper_download_model("large-v2-q5_0")
+#' whisper_download_model("large-v3-q5_0")
 #' }
 #' \dontshow{
 #' if(file.exists(path$file_model)) file.remove(path$file_model)
 #' }
-whisper_download_model <- function(x = c("tiny", "tiny.en", "base", "base.en", "small", "small.en", "medium", "medium.en", "large-v1", "large-v2", "large-v3", "large"),
+whisper_download_model <- function(x = c("tiny", "tiny.en", "base", "base.en", "small", "small.en", "medium", "medium.en", "large-v1", "large-v2", "large-v3", "large",
+                                         "tiny-q5_1", "tiny.en-q5_1", 
+                                         "base-q5_1", "base.en-q5_1", 
+                                         "small-q5_1", "small.en-q5_1", 
+                                         "medium-q5_1", "medium.en-q5_1", 
+                                         "large-v2-q5_0", "large-v3-q5_0"),
                                    model_dir = Sys.getenv("WHISPER_MODEL_DIR", unset = getwd()),
                                    repos = c("huggingface", "ggerganov"),
                                    version = c("1.5.4", "1.2.1"),