+++ NEWS.md
@@ -1,6 +1,12 @@
 ## CHANGES IN audio.whisper VERSION 0.3.2
 
 - Documentation of arguments in predict.whisper
+- Add option to download quantised models
+  - tiny-q5_1, tiny.en-q5_1
+  - base-q5_1, base.en-q5_1
+  - small-q5_1, small.en-q5_1
+  - medium-q5_1, medium.en-q5_1
+  - large-v2-q5_0 and large-v3-q5_0
 
 ## CHANGES IN audio.whisper VERSION 0.3.1